// src/interface/directorio.interface.tsx
// Generated by https://quicktype.io

export interface ResultDirectorios {
    ok:         boolean;
    msg:        string;
    directorio: Directorio[];
}

export interface Directorio {
    id:        number;
    nombre:    string;
    telefono:  string;
    direccion: string;
    lat:       number;
    lng:       number;
    logo:      string;
    imagen:    string;
    social:    string;
    estado:    number;
}